pyproject.toml
[project]
name = "epic-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.27.0",
    "bitsandbytes>=0.42.0",
    "chromadb>=0.4.24",
    "docling>=2.31.0",
    "fastapi>=0.109.2",
    "httpx==0.26.0",
    "langchain>=0.1.9",
    "langchain-chroma>=0.0.4",
    "langchain-community>=0.0.38",
    "langchain-huggingface>=0.0.2",
    "llama-cpp-python>=0.3.8",
    "loguru==0.7.2",
    "pypdf>=4.0.2",
    "pytest==8.0.1",
    "pytest-asyncio==0.23.5",
    "python-docx>=1.0.0",
    "python-dotenv==1.0.1",
    "python-multipart>=0.0.9",
    "requests>=2.30.0",
    "sentence-transformers>=2.5.1",
    "tqdm>=4.66.2",
    "transformers>=4.38.0",
    "uvicorn>=0.27.1",
]
